+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Changelog
 -   Added support for the `XY` (parameterized `iSWAP`) gate family in `Program`s
     and in `ISA`s (@ecpeterson, gh-1096, gh-1107, gh-1111).
 -   Removed the `tox.ini` and `readthedocs.yml` files (@karalekas, gh-1108).
+-   Type hints have been added to the `PauliSum` class (@rht, gh-1104).
 
 ### Bugfixes